requirement.txt
asgiref==3.5.0
Django==4.0.3
django-tinymce==3.4.0
django-versatileimagefield==2.2
Pillow==9.0.1
python-decouple==3.6
python-magic==0.4.25
python-magic-bin==0.4.14
sqlparse==0.4.2
tzdata==2022.1
